Agrego CRUD al carrito de compras

+++ fruteria/views.py
@@ -1,7 +1,10 @@
 from django.shortcuts import render, redirect
 from django.urls import reverse
+from django.contrib import messages
 from fruteria.models import Fruit , PurchaseCart
 from fruteria.forms import PurchaseCartFormulario
+import random
+import string
 
 def presentar_home(request):
     contexto = {}
@@ -43,11 +46,14 @@ def presentar_about(request):
     return http_responde
 
 
-def presentar_detail(request, id,mensaje=None):
+def presentar_detail(request, id):
     formulario =  PurchaseCartFormulario()
-    fruta = Fruit.objects.get(id = id)
+    fruit = Fruit.objects.get(id = id)
     atributo_url= request.path.split('/')[1]
-    contexto = {"fruta": fruta, "form":formulario, "mensaje":mensaje,"atributo_url":atributo_url}
+    contexto = {"fruta": fruit, "form":formulario, "atributo_url":atributo_url}
+    if atributo_url == 'cart':
+        quantity = PurchaseCart.objects.get(fruit=fruit, user= request.user, state = "en_carrito").quantity
+        contexto["quantity"] = quantity
     http_responde = render(
         request=request,
         template_name='fruteria/detalle.html',
@@ -66,7 +72,6 @@ def presentar_mensaje(request):
 
 
 def add_fruit(request,fruit_id):
-  
     if request.method == "POST":
         formulario = PurchaseCartFormulario(request.POST)
         if formulario.is_valid():
@@ -82,10 +87,11 @@ def add_fruit(request,fruit_id):
                 )
                 return http_responde
             else:
-                mensaje='Ya posee en su carrito esta fruta, puede ingresar al carrito para modificar la cantidadde su compra.'
-                url_fallida = reverse('detail', args=[fruit_id,mensaje]) 
+                messages.error(request, "Ya posee en su carrito esta fruta, puede ingresar al carrito para modificar la cantidad de kg que desea comprar.", extra_tags='tag1')
+                url_fallida = reverse('detail', args=[fruit_id]) 
                 return redirect(url_fallida)
         
+    
 def show_cart(request):
     carrito = PurchaseCart.objects.filter(user= request.user, state="en_carrito")
     total = sum([fruta.fruit.price*fruta.quantity for fruta in carrito])
@@ -97,17 +103,17 @@ def show_cart(request):
     )
     return http_responde
 
-def delete_fruit(request, fruit_id):
-    fruit = Fruit.objects.get(id=fruit_id)
-    cart_fruit = PurchaseCart.objects.get(fruit=fruit, user= request.user, state = "en_carrito")
-    cart_fruit.delete()
-    url = reverse('cart') 
-    return redirect(url)
 
+def delete_fruit(request, fruit_id):
+    if request.method == "POST":
+        fruit = Fruit.objects.get(id=fruit_id)
+        cart_fruit = PurchaseCart.objects.get(fruit=fruit, user= request.user, state = "en_carrito")
+        cart_fruit.delete()
+        url = reverse('cart') 
+        return redirect(url)
 
 
 def edit_fruit(request,fruit_id):
-  
     if request.method == "POST":
         formulario = PurchaseCartFormulario(request.POST)
         if formulario.is_valid():
@@ -118,3 +124,14 @@ def edit_fruit(request,fruit_id):
             purchase_cart.save()  # Lo guardan en la Base de datos
             url = reverse('cart') 
             return redirect(url)
+
+def get_purchase(request):
+    if request.method == "POST":
+        purchase_cart = PurchaseCart.objects.filter(user= request.user, state = "en_carrito")
+        code = ''.join(random.choices(string.ascii_uppercase + string.digits, k=15))
+        for purchase in purchase_cart:
+            purchase.state = "en_proceso_de_compra"
+            purchase.purchase_code = code
+            purchase.save()
+        url = reverse('cart') 
+        return redirect(url)
